pyproject.toml
[project]
name = "vecnote"
version = "0.1.0"
description = "a simple note taking app with small llm features"
authors = [{ name = "vemana arun sai", email = "arunvemana@outlook.com" }]
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.12.0",
]
[project.scripts]
vecnote = "VecNote.__main__:start"
[tool.uv]
package = true
